transform_gizmo_egui::config

Struct GizmoConfig

pub struct GizmoConfig {

    pub view_matrix: RowMatrix4<f64>,
    pub projection_matrix: RowMatrix4<f64>,
    pub viewport: Rect,
    pub modes: EnumSet<GizmoMode>,
    pub mode_override: Option<GizmoMode>,
    pub orientation: GizmoOrientation,
    pub pivot_point: TransformPivotPoint,
    pub snapping: bool,
    pub snap_angle: f32,
    pub snap_distance: f32,
    pub snap_scale: f32,
    pub visuals: GizmoVisuals,
    pub pixels_per_point: f32,

}

Configuration of a gizmo.

Defines how the gizmo is drawn to the screen and how it can be interacted with.

Fields
view_matrix: RowMatrix4<f64>

View matrix for the gizmo, aligning it with the cameraโs viewpoint.

projection_matrix: RowMatrix4<f64>

Projection matrix for the gizmo, determining how it is projected onto the screen.

viewport: Rect

Screen area where the gizmo is displayed.

modes: EnumSet<GizmoMode>

The gizmoโs operation modes.

mode_override: Option<GizmoMode>

If set, this mode is forced active and other modes are disabled

orientation: GizmoOrientation

Determines the gizmoโs orientation relative to global or local axes.

pivot_point: TransformPivotPoint

Pivot point for transformations

snapping: bool

Toggles snapping to predefined increments during transformations for precision.

snap_angle: f32

Angle increment for snapping rotations, in radians.

snap_distance: f32

Distance increment for snapping translations.

snap_scale: f32

Scale increment for snapping scalings.

visuals: GizmoVisuals

Visual settings for the gizmo, affecting appearance and visibility.

pixels_per_point: f32

Ratio of windowโs physical size to logical size.