transform_gizmo_egui::prelude

Struct Gizmo

pub struct Gizmo { /* private fields */ }

A 3D transformation gizmo.

Implementations

impl Gizmo

pub fn new(config: GizmoConfig) -> Gizmo

Creates a new gizmo from given configuration

pub fn config(&self) -> &GizmoConfig

Current configuration used by the gizmo.

pub fn update_config(&mut self, config: GizmoConfig)

Updates the configuration used by the gizmo.

pub fn is_focused(&self) -> bool

Was this gizmo focused after the latest Gizmo::update call.

pub fn update( &mut self, interaction: GizmoInteraction, targets: &[Transform], ) -> Option<(GizmoResult, Vec<Transform>)>

Updates the gizmo based on given interaction information.

Examples

let interaction = GizmoInteraction {
    cursor_pos,
    hovered,
    drag_started,
    dragging
};

if let Some((_result, new_transforms)) = gizmo.update(interaction, &transforms) {
                for (new_transform, transform) in
    // Update transforms
    new_transforms.iter().zip(&mut transforms)
    {
        *transform = *new_transform;
    }
}

Returns the result of the interaction with the updated transformation.

Some is returned when any of the subgizmos is being dragged, None otherwise.

pub fn draw(&self) -> GizmoDrawData

Return all the necessary data to draw the latest gizmo interaction.

The gizmo draw data consists of vertices in viewport coordinates.